fix(reader): preserve bridge bookmark timestamp contracts
Why:
- PR review found typed bridge DTOs widened bookmark timestamps from integer milliseconds to fractional JSON numbers.
- StudyPad document loading could also mark the controller as showing StudyPad before label serialization succeeded.

What Changed:
- Changed Bible and generic bookmark bridge DTO timestamp fields to integer milliseconds.
- Centralized Date-to-epoch-millis conversion in BibleReaderController and reused it for standard, StudyPad, and generic bookmark payloads.
- Validated StudyPad label DTOs before mutating StudyPad visibility state.
- Added raw bridge JSON assertions covering integer timestamp emission for bookmark update and StudyPad payloads.

Validation:
- git diff --check
- gitnexus_detect_changes scope all
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -scheme AndBible -destination 'platform=iOS Simulator,name=iPhone 16,OS=18.6' -only-testing:AndBibleTests/AndBibleTests/testReaderBookmarkBridgeUpdateEmitsTypedPayloadShape -only-testing:AndBibleTests/AndBibleTests/testReaderStudyPadDocumentBridgeEmissionUsesTypedNestedPayloads
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -scheme AndBible -destination 'platform=iOS Simulator,name=iPhone 16,OS=18.6' -only-testing:AndBibleTests/AndBibleTests

Impact:
- Restores Android and previous iOS bridge timestamp shape while preserving typed payload encoding.
- Reduces stale StudyPad state risk when labels disappear before serialization.

Refs:
- #176

+++ Sources/BibleView/Sources/BibleView/BridgeTypes.swift
@@ -280,8 +280,8 @@ public struct BibleBookmarkData: Codable, Sendable {
     public var bookName: String
     /// Compact module abbreviation.
     public var bookAbbreviation: String
-    /// Creation timestamp in milliseconds since 1970.
-    public var createdAt: Double // timestamp
+    /// Creation timestamp in integer milliseconds since 1970.
+    public var createdAt: Int // timestamp
     /// Short text excerpt shown in bookmark lists.
     public var text: String
     /// Full text payload available for notes and sharing.
@@ -290,8 +290,8 @@ public struct BibleBookmarkData: Codable, Sendable {
     public var bookmarkToLabels: [BookmarkToLabelData]
     /// Identifier of the primary label controlling styling precedence, or `nil` when unset.
     public var primaryLabelId: IdType?
-    /// Last modification timestamp in milliseconds since 1970.
-    public var lastUpdatedOn: Double
+    /// Last modification timestamp in integer milliseconds since 1970.
+    public var lastUpdatedOn: Int
     /// Optional user-authored note.
     public var notes: String?
     /// Precomputed note-presence flag for cheap client rendering.
@@ -329,12 +329,12 @@ public struct BibleBookmarkData: Codable, Sendable {
         bookInitials: String,
         bookName: String,
         bookAbbreviation: String,
-        createdAt: Double,
+        createdAt: Int,
         text: String,
         fullText: String,
         bookmarkToLabels: [BookmarkToLabelData],
         primaryLabelId: IdType?,
-        lastUpdatedOn: Double,
+        lastUpdatedOn: Int,
         notes: String?,
         hasNote: Bool,
         wholeVerse: Bool,
@@ -436,8 +436,8 @@ public struct GenericBookmarkData: Codable, Sendable {
     public var bookName: String
     /// Compact module abbreviation.
     public var bookAbbreviation: String
-    /// Creation timestamp in milliseconds since 1970.
-    public var createdAt: Double
+    /// Creation timestamp in integer milliseconds since 1970.
+    public var createdAt: Int
     /// Short text excerpt shown in bookmark lists.
     public var text: String
     /// Full text payload available for notes and sharing.
@@ -446,8 +446,8 @@ public struct GenericBookmarkData: Codable, Sendable {
     public var bookmarkToLabels: [BookmarkToLabelData]
     /// Identifier of the primary label controlling styling precedence, or `nil` when unset.
     public var primaryLabelId: IdType?
-    /// Last modification timestamp in milliseconds since 1970.
-    public var lastUpdatedOn: Double
+    /// Last modification timestamp in integer milliseconds since 1970.
+    public var lastUpdatedOn: Int
     /// Optional user-authored note.
     public var notes: String?
     /// Precomputed note-presence flag for cheap client rendering.
@@ -477,12 +477,12 @@ public struct GenericBookmarkData: Codable, Sendable {
         bookInitials: String,
         bookName: String,
         bookAbbreviation: String,
-        createdAt: Double,
+        createdAt: Int,
         text: String,
         fullText: String,
         bookmarkToLabels: [BookmarkToLabelData],
         primaryLabelId: IdType?,
-        lastUpdatedOn: Double,
+        lastUpdatedOn: Int,
         notes: String?,
         hasNote: Bool,
         wholeVerse: Bool,
